requirements.txt
apache-airflow~=3.0.3
pandas~=2.3.1
langchain~=0.3.26
python-dotenv~=1.1.1